refactor(gateway): split GatewayManager god object into manager/ submodules
manager.rs 5040 -> 133 LOC (struct + fields + mod decls + re-exports). Method
bodies moved verbatim into 13 responsibility-scoped impl blocks under manager/
(core, config_ops, pool_lifecycle, code_mode_runtime, code_mode_resolve,
persist, imports, import_matchers, virtual_servers, protected_routes,
oauth_resources, views) + tests/ split by concern. No file >500 LOC. Pure
structural move, zero behavior change: clippy -D warnings clean, fmt clean,
1611/1611 tests pass.
