+++ crates/lab/src/dispatch/gateway/manager/imports.rs
@@ -1,0 +1,336 @@
+//! Discovery import orchestration: auto-import, the pending-import queue, and
+//! import tombstone management.
+
+use crate::config::{UpstreamConfig, UpstreamImportTombstone};
+use crate::dispatch::error::ToolError;
+use crate::dispatch::gateway::config::{insert_upstream, write_gateway_config};
+use crate::dispatch::gateway::types::{
+    GatewayView, ImportErrorView, ImportResultView, ImportSkipReason, ImportSkipView,
+    ImportTombstoneView, PendingDiscoveryOutcome, PendingImportView,
+};
+use crate::dispatch::upstream::types::UpstreamRuntimeOwner;
+
+use super::GatewayManager;
+use super::import_matchers::{
+    ImportTombstoneSelector, discovered_is_tombstoned, partition_discovered_for_import,
+};
+
+fn pending_import_view(upstream: &UpstreamConfig) -> PendingImportView {
+    PendingImportView {
+        name: upstream.name.clone(),
+        url: upstream.url.clone(),
+        command: upstream.command.clone(),
+        source_client: upstream
+            .imported_from
+            .as_ref()
+            .map(|source| source.client.clone())
+            .unwrap_or_default(),
+        source_path: upstream
+            .imported_from
+            .as_ref()
+            .map(|source| source.path.clone())
+            .unwrap_or_default(),
+    }
+}
+
+fn import_tombstone_view(tombstone: &UpstreamImportTombstone) -> ImportTombstoneView {
+    ImportTombstoneView {
+        name: tombstone.name.clone(),
+        source_client: tombstone.imported_from.client.clone(),
+        source_path: tombstone.imported_from.path.clone(),
+        server_name: tombstone.imported_from.server_name.clone(),
+        transport_fingerprint: tombstone.imported_from.transport_fingerprint.clone(),
+        removed_at: tombstone.removed_at.clone(),
+    }
+}
+
+impl GatewayManager {
+    /// Import all externally configured MCP servers that are not already in the gateway config.
+    ///
+    /// Discovery-produced specs are intentionally persisted with `enabled = false`.
+    /// This makes newly found servers visible and editable in the gateway without
+    /// exposing them to downstream MCP clients until an operator enables them.
+    pub async fn auto_import_discovered_configs(&self) -> Result<ImportResultView, ToolError> {
+        let Some(home) = crate::dispatch::gateway::discovery::home_dir() else {
+            tracing::warn!(
+                surface = "dispatch",
+                service = "gateway",
+                action = "gateway.import",
+                event = "auto_import.skipped",
+                reason = "home_dir_unavailable",
+                "automatic MCP config import skipped"
+            );
+            return Ok(ImportResultView::default());
+        };
+
+        let discovered = tokio::task::spawn_blocking(move || {
+            crate::dispatch::gateway::discovery::discover_all(&home)
+        })
+        .await
+        .map_err(|e| ToolError::internal_message(format!("discovery task panicked: {e}")))?;
+
+        let cfg = self.current_config().await;
+        let (mut result, specs_to_add) = partition_discovered_for_import(&cfg, discovered);
+
+        if !specs_to_add.is_empty() {
+            let outcome = self
+                .batch_add(specs_to_add, Some("gateway.auto_import"), None)
+                .await?;
+            result.imported.extend(outcome.views);
+            for (name, err) in outcome.errors {
+                if matches!(err, ToolError::Conflict { .. }) {
+                    result.skipped.push(ImportSkipView {
+                        name,
+                        reason: ImportSkipReason::Conflict,
+                    });
+                } else {
+                    result.errors.push(ImportErrorView {
+                        name,
+                        message: err.to_string(),
+                    });
+                }
+            }
+        }
+
+        tracing::info!(
+            surface = "dispatch",
+            service = "gateway",
+            action = "gateway.import",
+            event = "auto_import.finish",
+            imported = result.imported.len(),
+            skipped = result.skipped.len(),
+            errors = result.errors.len(),
+            "automatic MCP config import finished"
+        );
+
+        Ok(result)
+    }
+
+    /// Scan external MCP configs and add newly discovered servers to the
+    /// `upstream_pending` queue without applying them.
+    pub async fn discover_into_pending(&self) -> Result<PendingDiscoveryOutcome, ToolError> {
+        let Some(home) = crate::dispatch::gateway::discovery::home_dir() else {
+            return Ok(PendingDiscoveryOutcome::default());
+        };
+
+        let discovered = tokio::task::spawn_blocking(move || {
+            crate::dispatch::gateway::discovery::discover_all(&home)
+        })
+        .await
+        .map_err(|e| ToolError::internal_message(format!("discovery task panicked: {e}")))?;
+
+        let _mutation_guard = self.config_mutation.lock().await;
+        let mut cfg = self.config.read().await.clone();
+
+        let mut queued = 0usize;
+        let mut skipped = 0usize;
+
+        for server in discovered {
+            // Skip tombstoned.
+            if discovered_is_tombstoned(&cfg, &server) {
+                skipped += 1;
+                continue;
+            }
+            // Skip already in upstreams.
+            if cfg.upstream.iter().any(|u| u.name == server.name) {
+                skipped += 1;
+                continue;
+            }
+            // Skip already pending.
+            if cfg.upstream_pending.iter().any(|u| u.name == server.name) {
+                skipped += 1;
+                continue;
+            }
+            cfg.upstream_pending.push(server.spec);
+            queued += 1;
+        }
+
+        if queued > 0 {
+            let path = self.path.clone();
+            let cfg_clone = cfg.clone();
+            tokio::task::spawn_blocking(move || write_gateway_config(&path, &cfg_clone))
+                .await
+                .map_err(|e| {
+                    ToolError::internal_message(format!("config write task failed: {e}"))
+                })??;
+            *self.config.write().await = cfg;
+        }
+
+        Ok(PendingDiscoveryOutcome { queued, skipped })
+    }
+
+    /// List the `upstream_pending` queue.
+    pub async fn list_pending_imports(&self) -> Vec<PendingImportView> {
+        let cfg = self.config.read().await;
+        cfg.upstream_pending
+            .iter()
+            .map(pending_import_view)
+            .collect()
+    }
+
+    /// Approve a pending import by name: move from `upstream_pending` into `upstream`
+    /// with `enabled = false` (same as auto-import — operator must explicitly enable).
+    pub async fn approve_pending_import(&self, name: &str) -> Result<PendingImportView, ToolError> {
+        let _mutation_guard = self.config_mutation.lock().await;
+        let mut cfg = self.config.read().await.clone();
+
+        let idx = cfg
+            .upstream_pending
+            .iter()
+            .position(|u| u.name == name)
+            .ok_or_else(|| ToolError::Sdk {
+                sdk_kind: "not_found".into(),
+                message: format!("pending import `{name}` not found"),
+            })?;
+
+        let mut spec = cfg.upstream_pending.remove(idx);
+        let view = pending_import_view(&spec);
+
+        spec.enabled = false;
+        cfg.upstream.push(spec);
+        let path = self.path.clone();
+        let cfg_clone = cfg.clone();
+        tokio::task::spawn_blocking(move || write_gateway_config(&path, &cfg_clone))
+            .await
+            .map_err(|e| ToolError::internal_message(format!("config write task failed: {e}")))??;
+        *self.config.write().await = cfg;
+
+        Ok(view)
+    }
+
+    /// Reject a pending import by name: tombstone it so it never re-appears.
+    pub async fn reject_pending_import(&self, name: &str) -> Result<PendingImportView, ToolError> {
+        let _mutation_guard = self.config_mutation.lock().await;
+        let mut cfg = self.config.read().await.clone();
+
+        let idx = cfg
+            .upstream_pending
+            .iter()
+            .position(|u| u.name == name)
+            .ok_or_else(|| ToolError::Sdk {
+                sdk_kind: "not_found".into(),
+                message: format!("pending import `{name}` not found"),
+            })?;
+
+        let spec = cfg.upstream_pending.remove(idx);
+        let view = pending_import_view(&spec);
+
+        // Create a tombstone so this server is never re-discovered.
+        if let Some(source) = spec.imported_from {
+            cfg.upstream_import_tombstones
+                .push(UpstreamImportTombstone::now(spec.name, source));
+        }
+
+        let path = self.path.clone();
+        let cfg_clone = cfg.clone();
+        tokio::task::spawn_blocking(move || write_gateway_config(&path, &cfg_clone))
+            .await
+            .map_err(|e| ToolError::internal_message(format!("config write task failed: {e}")))??;
+        *self.config.write().await = cfg;
+
+        Ok(view)
+    }
+
+    pub async fn list_import_tombstones(&self) -> Vec<ImportTombstoneView> {
+        let cfg = self.config.read().await;
+        cfg.upstream_import_tombstones
+            .iter()
+            .map(import_tombstone_view)
+            .collect()
+    }
+
+    pub async fn clear_import_tombstone(
+        &self,
+        selector: ImportTombstoneSelector,
+    ) -> Result<Vec<ImportTombstoneView>, ToolError> {
+        let _mutation_guard = self.config_mutation.lock().await;
+        let mut cfg = self.config.read().await.clone();
+        let before = cfg.upstream_import_tombstones.len();
+        cfg.upstream_import_tombstones
+            .retain(|tombstone| !selector.matches_tombstone(tombstone));
+        if cfg.upstream_import_tombstones.len() == before {
+            return Err(ToolError::Sdk {
+                sdk_kind: "not_found".to_string(),
+                message: format!("gateway import tombstone `{}` not found", selector.name),
+            });
+        }
+
+        let views: Vec<_> = cfg
+            .upstream_import_tombstones
+            .iter()
+            .map(import_tombstone_view)
+            .collect();
+        self.persist_config(cfg).await?;
+        Ok(views)
+    }
+
+    pub async fn restore_import_tombstone(
+        &self,
+        selector: ImportTombstoneSelector,
+        origin: Option<&str>,
+        owner: Option<UpstreamRuntimeOwner>,
+    ) -> Result<GatewayView, ToolError> {
+        let Some(home) = crate::dispatch::gateway::discovery::home_dir() else {
+            return Err(ToolError::internal_message(
+                "cannot determine home directory for MCP config restore",
+            ));
+        };
+        let discovered = tokio::task::spawn_blocking(move || {
+            crate::dispatch::gateway::discovery::discover_all(&home)
+        })
+        .await
+        .map_err(|e| ToolError::internal_message(format!("discovery task panicked: {e}")))?;
+
+        let _mutation_guard = self.config_mutation.lock().await;
+        let mut cfg = self.config.read().await.clone();
+        if !cfg
+            .upstream_import_tombstones
+            .iter()
+            .any(|tombstone| selector.matches_tombstone(tombstone))
+        {
+            return Err(ToolError::Sdk {
+                sdk_kind: "not_found".to_string(),
+                message: format!("gateway import tombstone `{}` not found", selector.name),
+            });
+        }
+
+        let Some(server) = discovered.into_iter().find(|server| {
+            selector.matches_discovered(server) && discovered_is_tombstoned(&cfg, server)
+        }) else {
+            return Err(ToolError::Sdk {
+                sdk_kind: "not_found".to_string(),
+                message: format!(
+                    "discovered MCP server `{}` no longer matches a tombstoned import",
+                    selector.name
+                ),
+            });
+        };
+
+        if cfg
+            .upstream
+            .iter()
+            .any(|upstream| upstream.name == server.spec.name)
+        {
+            cfg.upstream_import_tombstones
+                .retain(|tombstone| !selector.matches_tombstone(tombstone));
+            self.persist_config(cfg).await?;
+            return self.get(&server.spec.name).await;
+        }
+
+        let restored_name = server.spec.name.clone();
+        insert_upstream(&mut cfg, server.spec)?;
+        self.persist_config(cfg).await?;
+        let diff = self.reload_with_origin_unlocked(origin, owner).await?;
+        tracing::info!(
+            surface = "dispatch",
+            service = "gateway",
+            action = "gateway.import_tombstones.restore",
+            gateway = %restored_name,
+            tools_changed = diff.tools_changed,
+            resources_changed = diff.resources_changed,
+            prompts_changed = diff.prompts_changed,
+            "gateway import tombstone restored"
+        );
+        self.get(&restored_name).await
+    }
+}
